Video-Games-Sales-Analysis
The project is based on games selling throughout the world. The only purpose of the project is to get sales insights so that the management can take decision on publisher, genre and platform of the new product. Following are some of the outputs project is capable of:

1- Sales by Europe 2- Sales by North America 3- Sales by Japan 4- Global Sales 5- Sales by publisher 6- Sales by Genre 7- Sales by Platform 8- Sales by Year 9- Top N sales at any continent

Tools to be used:

1- Power BI 2- Mysql Workbench 3- Microsoft Excel

Techniques:

1- Data Extraction (ETL) 2- Data Cleansing 3- Data Analysis 4- Data Visualization